fix: RSEM-calculate-expression wrapper with FASTQ-files as input + Integration of test case (#970)
<!-- Ensure that the PR title follows conventional commit style (<type>:
<description>)-->
<!-- Possible types are here:
https://github.com/commitizen/conventional-commit-types/blob/master/index.json
-->

### Description

- Replaced deprecated argument for alignment inputs with appropriate
argument (`--bam`-> `--alignments`)
- Fixed wrong argument for FASTQ-input: Deleted `input_bam = False`
- Integrated test-case for FASTQ-input 
- Fixed environment-yaml-file: bowtie is needed for FASTQ-file input
- updated meta.yaml

### Notes
The original provided test-case from the authors for BAM-file inputs
includes only an empty BAM-file (except of the header). My test case
also includes reads, which do not map to the reference sequence (since
the provided reference sequence is too short anyways). So as I
understand, the purpose of these tests is just to ensure, whether the
pipeline is able to run through. However, to make it more robust
additional tests with real working examples would be needed.

Co-authored-by: Küchler, Oliver <oliver.kuechler@bih-charite.de>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>