Correctly handle non str index list
As reported in #522, currently this wrapper
has issues if `idx` is a sequence is it incorrectly
checked `input.index` to be as tring instead of `input.idx`.
This is fixed now.

While the discussin in #522 indicates, that there
is an underlying issue with the concept of `input.idx` being
as sequence of indices, this may need rework later.
Until all the details are decided, having a correct version
should be still valuable.