fix: try to get release-please.yml PRs to run tests (#3653)
<!-- Ensure that the PR title follows conventional commit style (<type>:
<description>)-->
<!-- Possible types are here:
https://github.com/commitizen/conventional-commit-types/blob/master/index.json
-->

This switches to the latest (and maintained) release-please GitHub
Action and tries to get the regular CI tests to run on release-please
pull requests via the `GITHUB_TOKEN`. Permissions are set as described
here:

https://github.com/googleapis/release-please-action?tab=readme-ov-file#workflow-permissions

But we might nevertheless have to [switch to a personal access
token](https://gist.github.com/dlaehnemann/0ab4f1e96660d81384405a04481dd31b)
for this, but this would be more hassle (and tied to a personal
account), so I would prefer if it works like this...

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] I confirm that I have followed the [documentation for contributing
to
`snakemake-wrappers`](https://snakemake-wrappers.readthedocs.io/en/stable/contributing.html).

While the contributions guidelines are more extensive, please
particularly ensure that:
* [x] `test.py` was updated to call any added or updated example rules
in a `Snakefile`
* [x] `input:` and `output:` file paths in the rules can be chosen
arbitrarily
* [x] wherever possible, command line arguments are inferred and set
automatically (e.g. based on file extensions in `input:` or `output:`)
* [x] temporary files are either written to a unique hidden folder in
the working directory, or (better) stored where the Python function
`tempfile.gettempdir()` points to
* [x] the `meta.yaml` contains a link to the documentation of the
respective tool or command under `url:`
* [x] conda environments use a minimal amount of channels and packages,
in recommended ordering


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Enhanced the release automation process with updated permission
settings and a refined configuration for authentication.
- Upgraded the underlying release tool to a newer version for improved
efficiency and security.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->